move auth to next middleware
